+++ src/index.test.ts
@@ -41,6 +41,7 @@ import {
   overwriteFile,
   getSolidDataset,
   createContainerInContainer,
+  deleteSolidDataset,
   saveSolidDatasetAt,
   saveSolidDatasetInContainer,
   GRANT_VC_URL_PARAM_NAME,
@@ -80,6 +81,7 @@ describe("Index exports", () => {
     expect(saveFileInContainer).toBeDefined();
     expect(createContainerInContainer).toBeDefined();
     expect(getSolidDataset).toBeDefined();
+    expect(deleteSolidDataset).toBeDefined();
     expect(saveSolidDatasetAt).toBeDefined();
     expect(saveSolidDatasetInContainer).toBeDefined();
     expect(GRANT_VC_URL_PARAM_NAME).toBeDefined();